Moved files to new folders, adjusted relative paths.
