+++ tests/unit/app.tests.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow, configure } from 'enzyme';
 import { expect } from 'chai';
 import Adapter from 'enzyme-adapter-react-16';
-import App from '../imports/components/app.js';
+import App from '../../imports/components/app.js';
 
 if (Meteor.isClient) {
   configure({ adapter: new Adapter() });
