+++ tests/unit/error.tests.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow, configure } from 'enzyme';
 import { expect } from 'chai';
 import Adapter from 'enzyme-adapter-react-16';
-import Error from '../imports/components/error.js';
+import Error from '../../imports/components/error.js';
 
 if (Meteor.isClient) {
   configure({ adapter: new Adapter() });
