+++ tests/unit/loading.tests.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow, configure } from 'enzyme';
 import { expect } from 'chai';
 import Adapter from 'enzyme-adapter-react-16';
-import Loading from '../imports/components/loading.js';
+import Loading from '../../imports/components/loading.js';
 
 if (Meteor.isClient) {
   configure({ adapter: new Adapter() });
